django_mongo_firebase/requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Deprecated==1.2.14
Django==4.2.7
djangorestframework==3.14.0
dnspython==2.4.2
firebase-admin==6.2.0
gcloud==0.18.3
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.4
jwcrypto==1.5.0
msgpack==1.0.7
oauth2client==4.1.3
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pymongo==4.6.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-jwt==4.0.0
pytz==2023.3.post1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
wrapt==1.15.0
